requirements.txt
# Frontend
streamlit==1.27.0

# Backend
fastapi==0.116.1
uvicorn==0.35.0

# LangChain / LangGraph
langchain==0.3.0
langgraph==0.6.6
langchain-groq==0.2.2

# Utilities
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
